+++ packages/EasyPipeline/src/Exceptions/EmptyMiddlewareListException.php
@@ -1,0 +1,11 @@
+<?php
+declare(strict_types=1);
+
+namespace StepTheFkUp\EasyPipeline\Exceptions;
+
+use StepTheFkUp\EasyPipeline\Interfaces\EasyPipelineExceptionInterface;
+
+final class EmptyMiddlewareListException extends \InvalidArgumentException implements EasyPipelineExceptionInterface
+{
+    // No body needed.
+}