+++ packages/EasyPipeline/src/Exceptions/EmptyPipelinesListException.php
@@ -1,0 +1,12 @@
+<?php
+declare(strict_types=1);
+
+namespace StepTheFkUp\EasyPipeline\Exceptions;
+
+use InvalidArgumentException;
+use StepTheFkUp\EasyPipeline\Interfaces\EasyPipelineExceptionInterface as ExceptionInterface;
+
+final class EmptyPipelinesListException extends InvalidArgumentException implements ExceptionInterface
+{
+    // No body needed.
+}