+++ packages/EasyPipeline/src/Exceptions/PipelineNotFoundException.php
@@ -1,0 +1,12 @@
+<?php
+declare(strict_types=1);
+
+namespace StepTheFkUp\EasyPipeline\Exceptions;
+
+use InvalidArgumentException;
+use StepTheFkUp\EasyPipeline\Interfaces\EasyPipelineExceptionInterface;
+
+final class PipelineNotFoundException extends InvalidArgumentException implements EasyPipelineExceptionInterface
+{
+    // No body needed.
+}