+++ packages/EasyRepository/tests/Bridge/Laravel/Stubs/Repository1Stub.php
@@ -1,0 +1,9 @@
+<?php
+declare(strict_types=1);
+
+namespace StepTheFkUp\EasyRepository\Tests\Bridge\Laravel\Stubs;
+
+final class Repository1Stub
+{
+
+}