+++ CHANGELOG.md
@@ -1,6 +1,29 @@
-# Changelog
+<!-- dumped content start -->
 
-<!-- changelog-linker -->
+## v0.3.1 - 2019-03-05
+
+- [#11] Improve Laravel service providers for DX, Thanks to [@natepage]
+
+### EasyPipeline
+
+- [#10] Update service provider to publish and merge config, Thanks to [@natepage]
+
+### EasyRepository
+
+- [#9] Update Doctrine ORM repository to define manager as entity manager an…, Thanks to [@natepage]
+
+<!-- dumped content end -->
+
+<!-- dumped content start -->
+
+## [v0.3.0] - 2019-03-02
+
+### EasyPipeline
+
+- [#8] Create package, Thanks to [@natepage]
+- [#7] Fix typo, Thanks to [@ttam]
+
+<!-- dumped content end -->
 
 <!-- dumped content start -->
 
@@ -54,3 +77,10 @@
 [v0.2.0]: https://github.com/StepTheFkUp/StepTheFkUp/compare/v0.1.1...v0.2.0
 [#6]: https://github.com/StepTheFkUp/StepTheFkUp/pull/6
 [v0.2.1]: https://github.com/StepTheFkUp/StepTheFkUp/compare/v0.2.0...v0.2.1
+[#8]: https://github.com/StepTheFkUp/StepTheFkUp/pull/8
+[#7]: https://github.com/StepTheFkUp/StepTheFkUp/pull/7
+[@ttam]: https://github.com/ttam
+[#11]: https://github.com/StepTheFkUp/StepTheFkUp/pull/11
+[#10]: https://github.com/StepTheFkUp/StepTheFkUp/pull/10
+[#9]: https://github.com/StepTheFkUp/StepTheFkUp/pull/9
+[v0.3.0]: https://github.com/StepTheFkUp/StepTheFkUp/compare/v0.2.1...v0.3.0